major edits to the about page
to address issue #2, I have edited the about page in  format so that it reflects the three goals: a) purpose, b) people, and c) expectations

it is my believe that this meets the minimum threshold to close #2.

one small thing: I do not full understand the 'layout:' and 'title' things at the top. I will need to build the site locally to find out if they are showing up on the page.